Refactor Element into its own module
Move the Element class into its own module named element. There may be
generic subclasses that make sense to group together into a module.
